Moving network attributes to the agent
