+++ android-agent/src/main/java/io/opentelemetry/android/features/networkattrs/CurrentNetwork.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.features.network;
+package io.opentelemetry.android.features.networkattrs;
 
 import android.os.Build;
 import androidx.annotation.Nullable;
