+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkApplicationListener.java
@@ -7,9 +7,9 @@ package io.opentelemetry.android.instrumentation.network;
 
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
 
-import io.opentelemetry.android.features.network.CurrentNetwork;
-import io.opentelemetry.android.features.network.CurrentNetworkProvider;
-import io.opentelemetry.android.features.network.NetworkChangeListener;
+import io.opentelemetry.android.features.networkattrs.CurrentNetwork;
+import io.opentelemetry.android.features.networkattrs.CurrentNetworkProvider;
+import io.opentelemetry.android.features.networkattrs.NetworkChangeListener;
 import io.opentelemetry.android.internal.services.applifecycle.ApplicationStateListener;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.context.Context;
