+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkChangeMonitor.java
@@ -8,8 +8,8 @@ package io.opentelemetry.android.instrumentation.network;
 import android.app.Application;
 import androidx.annotation.NonNull;
 import io.opentelemetry.android.OpenTelemetryRum;
-import io.opentelemetry.android.features.network.CurrentNetwork;
-import io.opentelemetry.android.features.network.CurrentNetworkProvider;
+import io.opentelemetry.android.features.networkattrs.CurrentNetwork;
+import io.opentelemetry.android.features.networkattrs.CurrentNetworkProvider;
 import io.opentelemetry.android.instrumentation.AndroidInstrumentation;
 import io.opentelemetry.android.instrumentation.common.InstrumentedApplication;
 import io.opentelemetry.android.internal.services.ServiceManager;
